+++ src/Tool/IntelliToolLine.cpp
@@ -52,7 +52,7 @@ void IntelliToolLine::onMouseMoved(int x, int y){
 				case LineStyle::DOTTED_LINE:
 						QPoint p1 =start.x() <= next.x() ? start : next;
 						QPoint p2 =start.x() < next.x() ? next : start;
-						int m = (float)(p2.y()-p1.y())/(float)(p2.x()-p1.x())+0.5f;
+                        int m = static_cast<int>((p2.y()-p1.y())/(p2.x()-p1.x())+0.5f);
 						int c = start.y()-start.x()*m;
 
 						break;
